+++ tests/integration/stream-zarr-v2-compressed-to-filesystem.cpp
@@ -59,19 +59,49 @@ setup()
 
     ZarrDimensionProperties* dim;
     dim = settings.dimensions;
-    *dim = DIM("t", ZarrDimensionType_Time, array_timepoints, chunk_timepoints, 0);
+    *dim = DIM("t",
+               ZarrDimensionType_Time,
+               array_timepoints,
+               chunk_timepoints,
+               0,
+               nullptr,
+               1.0);
 
     dim = settings.dimensions + 1;
-    *dim = DIM("c", ZarrDimensionType_Channel, array_channels, chunk_channels, 0);
+    *dim = DIM("c",
+               ZarrDimensionType_Channel,
+               array_channels,
+               chunk_channels,
+               0,
+               nullptr,
+               1.0);
 
     dim = settings.dimensions + 2;
-    *dim = DIM("z", ZarrDimensionType_Space, array_planes, chunk_planes, 0);
+    *dim = DIM("z",
+               ZarrDimensionType_Space,
+               array_planes,
+               chunk_planes,
+               0,
+               "millimeter",
+               1.4);
 
     dim = settings.dimensions + 3;
-    *dim = DIM("y", ZarrDimensionType_Space, array_height, chunk_height, 0);
+    *dim = DIM("y",
+               ZarrDimensionType_Space,
+               array_height,
+               chunk_height,
+               0,
+               "micrometer",
+               0.9);
 
     dim = settings.dimensions + 4;
-    *dim = DIM("x", ZarrDimensionType_Space, array_width, chunk_width, 0);
+    *dim = DIM("x",
+               ZarrDimensionType_Space,
+               array_width,
+               chunk_width,
+               0,
+               "micrometer",
+               0.9);
 
     auto* stream = ZarrStream_create(&settings);
     ZarrStreamSettings_destroy_dimension_array(&settings);
@@ -97,18 +127,29 @@ verify_base_metadata(const nlohmann::json& meta)
     type = axes[0]["type"];
     EXPECT(name == "t", "Expected name to be 't', but got '", name, "'");
     EXPECT(type == "time", "Expected type to be 'time', but got '", type, "'");
+    EXPECT(!axes[0].contains("unit"),
+           "Expected unit to be missing, got ",
+           axes[0]["unit"].get<std::string>());
 
     name = axes[1]["name"];
     type = axes[1]["type"];
     EXPECT(name == "c", "Expected name to be 'c', but got '", name, "'");
     EXPECT(
       type == "channel", "Expected type to be 'channel', but got '", type, "'");
+    EXPECT(!axes[1].contains("unit"),
+           "Expected unit to be missing, got ",
+           axes[1]["unit"].get<std::string>());
 
     name = axes[2]["name"];
     type = axes[2]["type"];
+    unit = axes[2]["unit"];
     EXPECT(name == "z", "Expected name to be 'z', but got '", name, "'");
     EXPECT(
       type == "space", "Expected type to be 'space', but got '", type, "'");
+    EXPECT(unit == "millimeter",
+           "Expected unit to be 'millimeter', but got '",
+           unit,
+           "'");
 
     name = axes[3]["name"];
     type = axes[3]["type"];
@@ -147,9 +188,9 @@ verify_base_metadata(const nlohmann::json& meta)
     EXPECT_EQ(size_t, scale.size(), 5);
     EXPECT_EQ(double, scale[0].get<double>(), 1.0);
     EXPECT_EQ(double, scale[1].get<double>(), 1.0);
-    EXPECT_EQ(double, scale[2].get<double>(), 1.0);
-    EXPECT_EQ(double, scale[3].get<double>(), 1.0);
-    EXPECT_EQ(double, scale[4].get<double>(), 1.0);
+    EXPECT_EQ(double, scale[2].get<double>(), 1.4);
+    EXPECT_EQ(double, scale[3].get<double>(), 0.9);
+    EXPECT_EQ(double, scale[4].get<double>(), 0.9);
 }
 
 void