+++ tests/integration/stream-zarr-v3-2d-multiscale-to-filesystem.cpp
@@ -75,22 +75,36 @@ setup()
                ZarrDimensionType_Time,
                array_timepoints,
                chunk_timepoints,
-               shard_timepoints);
+               shard_timepoints,
+               nullptr,
+               1.0);
 
     dim = settings.dimensions + 1;
     *dim = DIM("c",
                ZarrDimensionType_Channel,
                array_channels,
                chunk_channels,
-               shard_channels);
+               shard_channels,
+               nullptr,
+               1.0);
 
     dim = settings.dimensions + 2;
-    *dim = DIM(
-      "y", ZarrDimensionType_Space, array_height, chunk_height, shard_height);
+    *dim = DIM("y",
+               ZarrDimensionType_Space,
+               array_height,
+               chunk_height,
+               shard_height,
+               "micrometer",
+               0.9);
 
     dim = settings.dimensions + 3;
-    *dim =
-      DIM("x", ZarrDimensionType_Space, array_width, chunk_width, shard_width);
+    *dim = DIM("x",
+               ZarrDimensionType_Space,
+               array_width,
+               chunk_width,
+               shard_width,
+               "micrometer",
+               0.9);
 
     auto* stream = ZarrStream_create(&settings);
     ZarrStreamSettings_destroy_dimension_array(&settings);
@@ -128,12 +142,18 @@ verify_group_metadata(const nlohmann::json& meta)
     type = axes[0]["type"];
     EXPECT(name == "t", "Expected name to be 't', but got '", name, "'");
     EXPECT(type == "time", "Expected type to be 'time', but got '", type, "'");
+    EXPECT(!axes[0].contains("unit"),
+           "Expected unit to be missing, got ",
+           axes[0]["unit"].get<std::string>());
 
     name = axes[1]["name"];
     type = axes[1]["type"];
     EXPECT(name == "c", "Expected name to be 'c', but got '", name, "'");
     EXPECT(
       type == "channel", "Expected type to be 'channel', but got '", type, "'");
+    EXPECT(!axes[1].contains("unit"),
+           "Expected unit to be missing, got ",
+           axes[1]["unit"].get<std::string>());
 
     name = axes[2]["name"];
     type = axes[2]["type"];
@@ -178,10 +198,10 @@ verify_group_metadata(const nlohmann::json& meta)
 
         const auto scale = coordinate_transformations[0]["scale"];
         EXPECT_EQ(size_t, scale.size(), 4); // Now 4 scale factors
-        EXPECT_EQ(int, scale[0].get<double>(), 1.0);
-        EXPECT_EQ(int, scale[1].get<double>(), 1.0);
-        EXPECT_EQ(int, scale[2].get<double>(), std::pow(2, level));
-        EXPECT_EQ(int, scale[3].get<double>(), std::pow(2, level));
+        EXPECT_EQ(double, scale[0].get<double>(), 1.0);
+        EXPECT_EQ(double, scale[1].get<double>(), 1.0);
+        EXPECT_EQ(double, scale[2].get<double>(), std::pow(2, level) * 0.9);
+        EXPECT_EQ(double, scale[3].get<double>(), std::pow(2, level) * 0.9);
     }
 }