+++ tests/integration/stream-zarr-v3-compressed-to-s3.cpp
@@ -207,26 +207,45 @@ setup()
                ZarrDimensionType_Time,
                array_timepoints,
                chunk_timepoints,
-               shard_timepoints);
+               shard_timepoints,
+               nullptr,
+               1.0);
 
     dim = settings.dimensions + 1;
     *dim = DIM("c",
                ZarrDimensionType_Channel,
                array_channels,
                chunk_channels,
-               shard_channels);
+               shard_channels,
+               nullptr,
+               1.0);
 
     dim = settings.dimensions + 2;
-    *dim = DIM(
-      "z", ZarrDimensionType_Space, array_planes, chunk_planes, shard_planes);
+    *dim = DIM("z",
+               ZarrDimensionType_Space,
+               array_planes,
+               chunk_planes,
+               shard_planes,
+               "millimeter",
+               1.4);
 
     dim = settings.dimensions + 3;
-    *dim = DIM(
-      "y", ZarrDimensionType_Space, array_height, chunk_height, shard_height);
+    *dim = DIM("y",
+               ZarrDimensionType_Space,
+               array_height,
+               chunk_height,
+               shard_height,
+               "micrometer",
+               0.9);
 
     dim = settings.dimensions + 4;
-    *dim =
-      DIM("x", ZarrDimensionType_Space, array_width, chunk_width, shard_width);
+    *dim = DIM("x",
+               ZarrDimensionType_Space,
+               array_width,
+               chunk_width,
+               shard_width,
+               "micrometer",
+               0.9);
 
     auto* stream = ZarrStream_create(&settings);
     ZarrStreamSettings_destroy_dimension_array(&settings);
@@ -263,18 +282,29 @@ verify_group_metadata(const nlohmann::json& meta)
     type = axes[0]["type"];
     EXPECT(name == "t", "Expected name to be 't', but got '", name, "'");
     EXPECT(type == "time", "Expected type to be 'time', but got '", type, "'");
+    EXPECT(!axes[0].contains("unit"),
+           "Expected unit to be missing, got ",
+           axes[0]["unit"].get<std::string>());
 
     name = axes[1]["name"];
     type = axes[1]["type"];
     EXPECT(name == "c", "Expected name to be 'c', but got '", name, "'");
     EXPECT(
       type == "channel", "Expected type to be 'channel', but got '", type, "'");
+    EXPECT(!axes[1].contains("unit"),
+           "Expected unit to be missing, got ",
+           axes[1]["unit"].get<std::string>());
 
     name = axes[2]["name"];
     type = axes[2]["type"];
+    unit = axes[2]["unit"];
     EXPECT(name == "z", "Expected name to be 'z', but got '", name, "'");
     EXPECT(
       type == "space", "Expected type to be 'space', but got '", type, "'");
+    EXPECT(unit == "millimeter",
+           "Expected unit to be 'millimeter', but got '",
+           unit,
+           "'");
 
     name = axes[3]["name"];
     type = axes[3]["type"];
@@ -311,11 +341,11 @@ verify_group_metadata(const nlohmann::json& meta)
 
     const auto scale = coordinate_transformations["scale"];
     EXPECT_EQ(size_t, scale.size(), 5);
-    EXPECT_EQ(int, scale[0].get<double>(), 1.0);
-    EXPECT_EQ(int, scale[1].get<double>(), 1.0);
-    EXPECT_EQ(int, scale[2].get<double>(), 1.0);
-    EXPECT_EQ(int, scale[3].get<double>(), 1.0);
-    EXPECT_EQ(int, scale[4].get<double>(), 1.0);
+    EXPECT_EQ(double, scale[0].get<double>(), 1.0);
+    EXPECT_EQ(double, scale[1].get<double>(), 1.0);
+    EXPECT_EQ(double, scale[2].get<double>(), 1.4);
+    EXPECT_EQ(double, scale[3].get<double>(), 0.9);
+    EXPECT_EQ(double, scale[4].get<double>(), 0.9);
 }
 
 void
@@ -378,6 +408,25 @@ verify_array_metadata(const nlohmann::json& meta)
            "Expected shuffle to be 'shuffle', got ",
            blosc_config["shuffle"].get<std::string>());
     EXPECT_EQ(int, blosc_config["typesize"].get<int>(), 2);
+
+    const auto& dimension_names = meta["dimension_names"];
+    EXPECT_EQ(size_t, dimension_names.size(), 5);
+
+    EXPECT(dimension_names[0].get<std::string>() == "t",
+           "Expected first dimension name to be 't', got ",
+           dimension_names[0].get<std::string>());
+    EXPECT(dimension_names[1].get<std::string>() == "c",
+           "Expected second dimension name to be 'c', got ",
+           dimension_names[1].get<std::string>());
+    EXPECT(dimension_names[2].get<std::string>() == "z",
+           "Expected third dimension name to be 'z', got ",
+           dimension_names[2].get<std::string>());
+    EXPECT(dimension_names[3].get<std::string>() == "y",
+           "Expected fourth dimension name to be 'y', got ",
+           dimension_names[3].get<std::string>());
+    EXPECT(dimension_names[4].get<std::string>() == "x",
+           "Expected fifth dimension name to be 'x', got ",
+           dimension_names[4].get<std::string>());
 }
 
 void